Added features for user and developer experience
Cleaned up the code to make adjustments easier to make, whilst also improving performance.  However multiple bugs were discovered, such as dragging a group of nodes will not be consistent when having to change the node dragged multiple times and multiple nodes not being possible to undo for the make operation, the latter was fixed.

Features adjusted:
    Adjusted create nodes to have an assigned keybind.  Insert.
    Selected nodes will now be stored, to allow for slightly faster access
    All nodes will now be retrievable in GroupSelector, speeds up loops.
    Cleaned up node, menuBar, toolBar, editPopupMenu, connector, and actionManager using selectedNodes and nodes.
    Added ability to select all nodes in menuBar, CTRL-A.
    Multiple node actions can now be undone/redone.
    Node connections are more seemless, but not completely finished.